Fix FM rf receive issue

+++ FMCTCSSRX.cpp
@@ -98,7 +98,7 @@ CFMCTCSSRX::~CFMCTCSSRX()
 {
 }
 
-uint8_t CFMCTCSSRX::setParams(uint8_t frequency, uint8_t highThreshold, uint8_t lowThreshold)
+uint8_t CFMCTCSSRX::setParams(uint8_t frequency, uint16_t highThreshold, uint16_t lowThreshold)
 {
   m_coeffDivTwo = 0;
